feat: color updates, partial character grid refactor (#741)
* feat: color updates

* refactor: partially extract characterGrid, fix mid-optimizer rendering lag

* feat: preload home i18n file to remove suspense loader

* feat: add suspense to character preview

* fix: getting started link
