+++ src/lib/i18n/i18n.ts
@@ -95,6 +95,7 @@ void i18next
       'optimizerTab',
       'notifications',
       'conditionals',
+      'hometab',
     ],
     defaultNS: 'common',
     fallbackNS: ['common', 'gameData'],
