+++ src/lib/tabs/Tabs.tsx
@@ -13,7 +13,7 @@ import RelicsTab from 'lib/tabs/tabRelics/RelicsTab'
 import RelicScorerTab from 'lib/tabs/tabShowcase/RelicScorerTab'
 import WebgpuTab from 'lib/tabs/tabWebgpu/WebgpuTab'
 import { WorkerPool } from 'lib/worker/workerPool'
-import React, { ReactElement, Suspense, useEffect } from 'react'
+import React, { ReactElement, useEffect } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
 import { Form } from 'types/form'
 
@@ -30,6 +30,7 @@ let optimizerInitialized = false
 const Tabs = () => {
   const activeKey = window.store((s) => s.activeKey).split('?')[0]
 
+  const homeTab = React.useMemo(() => <HomeTab/>, [])
   const optimizerTab = React.useMemo(() => <OptimizerTab/>, [])
   const characterTab = React.useMemo(() => <CharacterTab/>, [])
   const relicsTab = React.useMemo(() => <RelicsTab/>, [])
@@ -38,7 +39,6 @@ const Tabs = () => {
   const changelogTab = React.useMemo(() => <ChangelogTab/>, [])
   const webgpuTab = React.useMemo(() => <WebgpuTab/>, [])
   const metadataTab = React.useMemo(() => <MetadataTab/>, [])
-  const homeTab = React.useMemo(() => <Suspense><HomeTab/></Suspense>, [])
 
   useEffect(() => {
     let route = PageToRoute[activeKey] || PageToRoute[AppPages.OPTIMIZER]
